update conftest, make publisher subprocess a daemon
